+++ benchmarks/4_nested_loops.R
@@ -1,13 +1,13 @@
 # Benchmark 4: Nested loops (1M iterations)
 # Nested loop that runs 1M times total
 
-outer: int <- 1000
-inner: int <- 1000
-count: int <- 0
+outer: double <- 1000
+inner: double <- 1000
+count: double <- 0
 
-i: int <- 1
+i: double <- 1
 while (i <= outer) {
-  j: int <- 1
+  j: double <- 1
   while (j <= inner) {
     count <- count + 1
     j <- j + 1
