+++ data/control_flow/basic_if_else.R
@@ -1,4 +1,4 @@
-if_else_func <- function(): int {
+if_else_func <- function(): double {
     x <- 10
     y <- 0
     if (x > 12)
