+++ data/control_flow/basic_loop.R
@@ -1,4 +1,4 @@
-basic_loop <- function(): int {
+basic_loop <- function(): double {
     sum <- 0
     for(i in 1:5) {
         sum <- sum + i
