+++ data/control_flow/if_expr.R
@@ -1,10 +1,10 @@
 # Test if expression with simple values
-x: int <- 15
-my_val: int <- if(x > 10) 1 else 2
+x: double <- 15
+my_val: double <- if(x > 10) 1 else 2
 print(my_val)
 
 # Test if expression with blocks
-y: int <- if(x < 10) {
+y: double <- if(x < 10) {
     print(0)
     100
 } else {
